Refactor character-related components and add combat features
Updated character features to use more specific data structures and refined feature management. Added "Use for Martial Check" functionality across character skills and integrated new combat-specific attributes and UI elements for better character management and gameplay experience.
