+++ Src/SwDc20.WebBlazor/Pages/Characters/CharacterSheet.razor
@@ -50,7 +50,7 @@ else
 
         <h5>Class Features</h5>
         <ul>
-            @foreach (var feature in characterWrapper.Document.ClassFeatures)
+            @foreach (var feature in characterWrapper.Document.Features)
             {
                 <li>@feature</li>
             }
