+++ Src/SwDc20.WebBlazor/Pages/Characters/Characters.razor
@@ -6,6 +6,7 @@
 @inject IModalService ModalService
 @inject NavigationManager NavigationManager
 @inject CurrentCharacterService CurrentCharacterService
+@inject GamesService GamesService
 
 <h3>Characters</h3>
 
@@ -19,20 +20,54 @@ else
         <button class="btn btn-primary" @onclick="CreateNewCharacter">Create New Character</button>
     </div>
 
-    <div class="row">
+    <div class="row row-cols-auto">
         @foreach (var characterWrapper in characters)
         {
             <div class="col-md-4 mb-3">
                 <div class="card">
                     <div class="card-body">
-                        <h5 class="card-title">@characterWrapper.Document.Name</h5>
-                        <p class="card-text">Level: @characterWrapper.Document.Level</p>
-                        <p class="card-text">Class: @characterWrapper.Document.Class</p>
+                        <div class="row">
+                            <div class="col-6">
+                                <h5 class="card-title">@characterWrapper.Document.Name</h5>
+                                <p class="card-text">Level: @characterWrapper.Document.Level</p>
+                                <p class="card-text">Class: @characterWrapper.Document.Class</p>
+                                <p class="card-text">Sub Class: @characterWrapper.Document.Subclass</p>
+                                <p class="card-text">Ancestry: @characterWrapper.Document.Ancestry</p>
+                            </div>
+                            <div class="col-6">
+                                @if (!string.IsNullOrEmpty(characterWrapper.Document.ImageUrl))
+                                {
+                                    <img src="@characterWrapper.Document.ImageUrl" alt="Character Image" style="max-width: 200px; max-height: 200px;"/>
+                                }
+                            </div>
+                        </div>
+                        
                         <button class="btn btn-success mr-2" @onclick="() => PlayCharacter(characterWrapper.Document)">Play</button>
                         <button class="btn btn-primary mr-2" @onclick="() => EditCharacter(characterWrapper.Document.Id)">Edit</button>
                         <button class="btn btn-danger" @onclick="() => DeleteCharacter(characterWrapper.Document.Id)">Delete</button>
+                        
                     </div>
                 </div>
+                @* This might be something to handle to allow players to directly play
+                    but I need to leverage signal to allow for eventing between players *@
+                @* <div class="card-footer"> *@
+                @*     <table class="table table-sm "> *@
+                @*         <thead> *@
+                @*             <tr> *@
+                @*                 <th>Game Name</th> *@
+                @*                 <th>Level</th> *@
+                @*             </tr> *@
+                @*         </thead> *@
+                @*         @foreach (var game in currentCharacters) *@
+                @*         { *@
+                @* *@
+                @*             <tr> *@
+                @*                 <td>@game.</td> *@
+                @*                 <td>@game.Level</td> *@
+                @*             </tr> *@
+                @*         } *@
+                @*     </table> *@
+                @* </div> *@
             </div>
         }
     </div>
@@ -41,6 +76,7 @@ else
 @code {
     private List<DocumentWrapper<Character>> characters;
     private CurrentCharacter? currentCharacter;
+    // private List<Game> currentGamesWithCharacters;
     private ModalOptions modalOptions = new ModalOptions()
     {
         Size = ModalSize.ExtraLarge,
@@ -51,6 +87,7 @@ else
     {
         characters = await CharacterService.GetCharactersAsync();
         currentCharacter = await CurrentCharacterService.GetCurrentCharacterAsync();
+        // currentGamesWithCharacters = await GamesService.GetGamesWithCharactersAsync(currentCharacter.Character.Id);
     }
 
     private async Task CreateNewCharacter()
