+++ Src/SwDc20.WebBlazor/Pages/Characters/Modals/FeatureEditForm.razor
@@ -1,36 +1,36 @@
 ﻿@inject IModalService ModalService
 
-<EditForm Model="@Trade" OnValidSubmit="SaveFeature">
+<EditForm Model="@Feature" OnValidSubmit="SaveFeature">
     <DataAnnotationsValidator />
     <ValidationSummary />
 
     <div class="form-group">
         <label for="name">Name</label>
-        <InputText id="name" class="form-control" @bind-Value="Trade.Name" />
+        <InputText id="name" class="form-control" @bind-Value="Feature.Name" />
     </div>
     
     <div class="form-group">
         <label for="description">Description</label>
-        <InputText id="description" class="form-control" @bind-Value="Trade.Description" />
+        <InputTextArea id="description" class="form-control" @bind-Value="Feature.Description" />
     </div>
 
     <div class="form-group">
-        <label for="bonus">Bonus</label>
-        <InputNumber id="bonus" class="form-control" @bind-Value="Trade.Bonus" />
+        <label for="bonus">Tag</label>
+        <InputText id="bonus" class="form-control" @bind-Value="Feature.Tag" />
     </div>
 
-    <button type="submit" class="btn btn-primary">Save Trade</button>
+    <button type="submit" class="btn btn-primary">Save Feature</button>
     <button type="button" class="btn btn-secondary" @onclick="Cancel">Cancel</button>
 </EditForm>
 
 @code {
     [CascadingParameter] BlazoredModalInstance BlazoredModal { get; set; }
 
-    [Parameter] public Trade Trade { get; set; }
+    [Parameter] public CharacterFeature Feature { get; set; }
 
     private async Task SaveFeature()
     {
-        await BlazoredModal.CloseAsync(ModalResult.Ok(Trade));
+        await BlazoredModal.CloseAsync(ModalResult.Ok(Feature));
     }
 
     private async Task Cancel()
