+++ Src/SwDc20.WebBlazor/Pages/Skills/SkillModal.razor
@@ -33,6 +33,11 @@
                         <label for="tag">Tag</label>
                         <InputText id="tag" @bind-Value="Skill.Tag" class="form-control" />
                     </div>
+                    
+                    <div class="form-group">
+                        <InputCheckbox id="UseForMartialCheck" @bind-Value="Skill.UseForMartialCheck"></InputCheckbox>
+                        <label for="UseForMartialCheck">Use for Martial Check</label>
+                    </div>
 
                     <button type="submit" class="btn btn-primary">Save</button>
                 </EditForm>
